Moving the code that don't use the package protected class from spark streaming from org.apache.spark -> io.radanalytics
